Moved dataset schemas to the separate component
