+++ src/app/query/global-query/search-and-schemas-section/search-and-schemas-section.component.ts
@@ -1,0 +1,130 @@
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, inject, OnInit } from "@angular/core";
+import { GlobalQuerySearchItem } from "../global-query.model";
+import { DatasetSubscriptionsService } from "src/app/dataset-view/dataset.subscriptions.service";
+import { BaseComponent } from "src/app/common/base.component";
+import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
+import { combineLatest, debounceTime, distinctUntilChanged, map, Observable, OperatorFunction, switchMap } from "rxjs";
+import { DatasetBasicsFragment, GetDatasetSchemaQuery } from "src/app/api/kamu.graphql.interface";
+import { parseCurrentSchema } from "src/app/common/app.helpers";
+import { NgbTypeaheadSelectItemEvent } from "@ng-bootstrap/ng-bootstrap";
+import { MaybeNull } from "src/app/common/app.types";
+import { DatasetSchema } from "src/app/interface/dataset.interface";
+import { DatasetAutocompleteItem, TypeNames } from "src/app/interface/search.interface";
+import { SearchApi } from "src/app/api/search.api";
+import { DatasetService } from "src/app/dataset-view/dataset.service";
+import AppValues from "src/app/common/app.values";
+
+@Component({
+    selector: "app-search-and-schemas-section",
+    templateUrl: "./search-and-schemas-section.component.html",
+    styleUrls: ["./search-and-schemas-section.component.scss"],
+    changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class SearchAndSchemasSectionComponent extends BaseComponent implements OnInit {
+    private searchResult: GlobalQuerySearchItem[] = [];
+    public inputDatasets = new Set<string>();
+    public searchDataset = "";
+    private readonly delayTime: number = AppValues.SHORT_DELAY_MS;
+
+    private datasetSubsService = inject(DatasetSubscriptionsService);
+    private cdr = inject(ChangeDetectorRef);
+    private datasetService = inject(DatasetService);
+    private appSearchAPI = inject(SearchApi);
+
+    public ngOnInit(): void {
+        this.datasetSubsService.emitInvolvedSqlDatasetsId([]);
+        this.datasetSubsService.involvedSqlDatasetsIdChanges
+            .pipe(
+                switchMap((ids: string[]) => {
+                    const schemaResponses = ids.map((id: string) => this.datasetService.requestDatasetSchema(id));
+                    return this.processSchemaResponses(schemaResponses);
+                }),
+                takeUntilDestroyed(this.destroyRef),
+            )
+            .subscribe((data: GlobalQuerySearchItem[]) => {
+                this.searchResult = data;
+                this.cdr.detectChanges();
+            });
+    }
+
+    private processSchemaResponses(
+        schemaResponses: Observable<GetDatasetSchemaQuery>[],
+    ): Observable<GlobalQuerySearchItem[]> {
+        return combineLatest(schemaResponses).pipe(
+            map((datasets: GetDatasetSchemaQuery[]) => {
+                const result = datasets.map((item: GetDatasetSchemaQuery) => {
+                    const datasetAlias = item.datasets.byId?.alias;
+                    const schema = parseCurrentSchema(item.datasets.byId?.metadata.currentSchema);
+                    return { datasetAlias, schema } as GlobalQuerySearchItem;
+                });
+                return result;
+            }),
+        );
+    }
+
+    public removeDataset(datasetAlias: string): void {
+        this.searchResult = this.searchResult.filter(
+            (item: GlobalQuerySearchItem) => item.datasetAlias !== datasetAlias,
+        );
+        this.inputDatasets.forEach((item) => {
+            if (item.includes(datasetAlias)) {
+                this.inputDatasets.delete(item);
+            }
+        });
+    }
+
+    public search: OperatorFunction<string, readonly DatasetAutocompleteItem[]> = (text$: Observable<string>) => {
+        return text$.pipe(
+            debounceTime(this.delayTime),
+            distinctUntilChanged(),
+            switchMap((term: string) => this.appSearchAPI.autocompleteDatasetSearch(term)),
+        );
+    };
+
+    public formatter(x: DatasetAutocompleteItem | string): string {
+        return typeof x !== "string" ? x.dataset.name : x;
+    }
+
+    public onSelectItem(event: NgbTypeaheadSelectItemEvent): void {
+        const value = event.item as DatasetAutocompleteItem;
+        const id = value.dataset.id;
+        const name = value.dataset.name;
+        const inputDataset = JSON.stringify({
+            datasetRef: id,
+            alias: name,
+        });
+        if (value.__typename !== TypeNames.allDataType && !this.inputDatasets.has(inputDataset)) {
+            this.inputDatasets.add(inputDataset);
+            this.requestDatasetSchemaById(id);
+        }
+    }
+
+    public get sortedSearchResult(): GlobalQuerySearchItem[] {
+        return this.searchResult.sort((a: GlobalQuerySearchItem, b: GlobalQuerySearchItem) => {
+            if (a.datasetAlias < b.datasetAlias) return -1;
+            if (a.datasetAlias > b.datasetAlias) return 1;
+            return 0;
+        });
+    }
+
+    private requestDatasetSchemaById(id: string): void {
+        this.datasetService
+            .requestDatasetSchema(id)
+            .pipe(takeUntilDestroyed(this.destroyRef))
+            .subscribe((data: GetDatasetSchemaQuery) => {
+                if (data.datasets.byId) {
+                    const datasetAlias = (data.datasets.byId as DatasetBasicsFragment).alias;
+                    const schema: MaybeNull<DatasetSchema> = parseCurrentSchema(
+                        data.datasets.byId.metadata.currentSchema,
+                    );
+
+                    this.searchResult = [...this.searchResult, { datasetAlias, schema }];
+                    this.cdr.detectChanges();
+                }
+            });
+    }
+
+    public clearSearch(): void {
+        this.searchDataset = "";
+    }
+}
