+++ src/app/shared/editor/components/sql-editor/sql-editor.component.scss
@@ -4,7 +4,6 @@
     position: relative;
     border-radius: 3px;
     border: 1px solid $app-color-border-default;
-    z-index: 9999;
 }
 
 .cell-border-bottom {
@@ -28,6 +27,5 @@
     top: 20px;
     left: 65px;
     background: none;
-    z-index: 1;
     opacity: 0.5;
 }
